requirements.txt
django==4.2
uwsgi==2.0.28
psycopg2==2.9.10
Pillow==11.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
django-redis==5.4.0
